Commit for Nick to look into DB Driver not being found.
Attempting to process the file and get events in the database. Will need
to address the timestamp to be put into the DB correctly - currently
storing it as a string.

Also cleaned up a lot of the warnings and resolved errors.

Added my run.bat script which facilitates the ant clean, run, copy .chst
file and run again cycle for testing.